git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/trunk@2163 494c09c7-a67d-49d4-9938-151fb1965be8

+++ RobWorkSim/src/rwsim/util/FiniteStateSampler.cpp
@@ -41,7 +41,7 @@ bool FiniteStateSampler::sample(rw::kinematics::State& state){
 			case(ORDERED_SAMPLING):{
 				state = _states[_cidx];
 				_n--;
-				if(_states.size()==_cidx){
+				if(_states.size()==(size_t)_cidx){
 					_cidx=0;
 				}
 				break;
